feat: remove useless recipe descriptions
All these separate descriptions for recipes are cute, but not what
players are used to. Describe the product, not the process.

For all recipes that produce one item, Factorio will default to the
name and description for that item. For recipes with catalysts, copy
the name and description of the main item for the recipe.

The exceptions are rock-knapping and wood-stick-sharpen: keep separate
recipe descriptions for these that explain that ingredients may be
destroyed (that's why there's probabilities on the results).