feat: implemented hammering mechanic, added anvil
To make the API work, prevent the anvil from being opened. Instead,
players should insert copper bloom and hammer. The hammer is
automatically returned to the player and crafting progress raised. Once
complete, the player can retrieve the crafted crude copper. Also added a
message displayed when the player tries to open the anvil to make sure
the mechanic is properly explained.