feat: switch end goal from copper axe to iron axe.
The copper axe as end goal never made much sense: you'd immediately want
to replace it with an iron axe, which has the same capabilities but more
hitpoints.

In stead of malachite from copper patches, the stone axe now gives
magnetite flakes from iron patches. The rest of the process is the same,
just with different names.