feat: introduce digging mechanic
To dig a mixing pit, hold a shovel in your hand and drop it over the
place to dig the pit. Since this mechanic is perhaps unexpected, an
uncraftable recipe remains in the GUI, with the description explaining
how to dig a mixing pit without crafting. The stone-shovel description
also explains this.

Since digging the pits is now free, they cannot be picked up anymore,
only removed.

One of the mining sounds was re-used as the sound the shovel makes when
in use.