feat: change from stone hammer to copper hammer
The hammer is made from crude copper, made directly from smelted
malachite flakes that you get from mining copper patches with a stone
axe. Copper doesn't need to be hammered.

The energy usage of the kiln has been lowered so no additional batches
of charcoal are needed to smelt both the required copper and iron.

This way, mining from stone, iron and copper patches is incorporated
in the playthrough. It also increases uses of the kiln (#17).