feat: switch back to manual resource spawning
Changed the spawning algorythm somewhat: first choose an area to spawn
in that is preferably completely clear, and then spawn a semi-random
amount in random spots within that area. This way the clay and sand
patches are grouped together more, which looks better and is easier to
find and mine.